+++ showcase/integrations/langgraph-python/src/agents/beautiful_chat.py
@@ -101,8 +101,6 @@ def query_data(query: str):
     """
     Query the database, takes natural language. Always call before showing a chart or graph.
     """
-    import time
-    print(f"[A2UI-DEBUG] query_data called: query='{query[:60]}' at {time.strftime('%H:%M:%S')}")
     return _cached_data
 
 
@@ -110,16 +108,11 @@ def query_data(query: str):
 
 CATALOG_ID = "copilotkit://app-dashboard-catalog"
 SURFACE_ID = "flight-search-results"
+FLIGHT_SCHEMA = a2ui.load_schema(_DATA_DIR / "schemas" / "flight_schema.json")
 
 
-class Flight(TypedDict, total=False):
-    # Only `airline`-through-`price` are read by `_build_flight_components`.
-    # All fields marked optional (`total=False`) so the LLM (or aimock fixture)
-    # can omit auxiliary fields like `id` / `statusIcon` without tripping
-    # langchain's tool-arg validation. Previously these were required and any
-    # missing field surfaced as `Error invoking tool 'search_flights' with
-    # kwargs ... flights.N.id: Field required` — the agent treated the error
-    # string as the tool result and the surface never rendered.
+class Flight(TypedDict):
+    id: str
     airline: str
     airlineLogo: str
     flightNumber: str
@@ -130,48 +123,12 @@ class Flight(TypedDict, total=False):
     arrivalTime: str
     duration: str
     status: str
+    statusIcon: str
     price: str
 
 
-def _build_flight_components(flights: list[dict]) -> list[dict]:
-    """Build a flat A2UI component tree with one literal FlightCard per flight.
-
-    Avoids the structural-children template form (Row.children = { componentId,
-    path }), which the GenericBinder only expands correctly for components whose
-    schema declares STRUCTURAL children — sibling demos work because their
-    schemas use literal-string-array children. Inlining the values per-flight
-    sidesteps the template path entirely and renders identically.
-    """
-    flight_card_ids: list[str] = []
-    components: list[dict] = []
-    for index, flight in enumerate(flights):
-        card_id = f"flight-card-{index}"
-        flight_card_ids.append(card_id)
-        components.append({
-            "id": card_id,
-            "component": "FlightCard",
-            "airline": flight["airline"],
-            "airlineLogo": flight["airlineLogo"],
-            "flightNumber": flight["flightNumber"],
-            "origin": flight["origin"],
-            "destination": flight["destination"],
-            "date": flight["date"],
-            "departureTime": flight["departureTime"],
-            "arrivalTime": flight["arrivalTime"],
-            "duration": flight["duration"],
-            "status": flight["status"],
-            "price": flight["price"],
-        })
-    root: dict = {
-        "id": "root",
-        "component": "Row",
-        "children": flight_card_ids,
-    }
-    return [root, *components]
-
-
 @tool
-def search_flights(flights: list[dict]) -> str:
+def search_flights(flights: list[Flight]) -> str:
     """Search for flights and display the results as rich cards. Return exactly 2 flights.
 
     Each flight must have: id, airline (e.g. "United Airlines"),
@@ -192,7 +149,8 @@ def search_flights(flights: list[dict]) -> str:
     return a2ui.render(
         operations=[
             a2ui.create_surface(SURFACE_ID, catalog_id=CATALOG_ID),
-            a2ui.update_components(SURFACE_ID, _build_flight_components(flights)),
+            a2ui.update_components(SURFACE_ID, FLIGHT_SCHEMA),
+            a2ui.update_data_model(SURFACE_ID, {"flights": flights}),
         ],
     )
 
@@ -229,12 +187,7 @@ def generate_a2ui(runtime: ToolRuntime[Any]) -> str:
     A secondary LLM designs the UI schema and data. The result is
     returned as an a2ui_operations container for the middleware to detect.
     """
-    import time
-    t0 = time.time()
-    print(f"[A2UI-DEBUG] generate_a2ui STARTED at t=0")
-
     messages = runtime.state["messages"][:-1]
-    print(f"[A2UI-DEBUG]   messages count: {len(messages)}")
 
     # Get context entries from copilotkit state (catalog capabilities + component schema)
     context_entries = runtime.state.get("copilotkit", {}).get("context", [])
@@ -242,7 +195,6 @@ def generate_a2ui(runtime: ToolRuntime[Any]) -> str:
         entry.get("value", "") for entry in context_entries
         if isinstance(entry, dict) and entry.get("value")
     )
-    print(f"[A2UI-DEBUG]   context entries: {len(context_entries)}, context_text_len: {len(context_text)}")
 
     prompt = context_text
 
@@ -252,15 +204,11 @@ def generate_a2ui(runtime: ToolRuntime[Any]) -> str:
         tool_choice="render_a2ui",
     )
 
-    print(f"[A2UI-DEBUG]   calling secondary LLM at t={time.time()-t0:.1f}s")
     response = model_with_tool.invoke(
         [SystemMessage(content=prompt), *messages],
     )
-    print(f"[A2UI-RESPONSE] {response}")
-    print(f"[A2UI-DEBUG]   secondary LLM responded at t={time.time()-t0:.1f}s")
 
     if not response.tool_calls:
-        print(f"[A2UI-DEBUG]   ERROR: no tool calls in response")
         return json.dumps({"error": "LLM did not call render_a2ui"})
 
     tool_call = response.tool_calls[0]
@@ -270,7 +218,6 @@ def generate_a2ui(runtime: ToolRuntime[Any]) -> str:
     catalog_id = args.get("catalogId", CUSTOM_CATALOG_ID)
     components = args.get("components", [])
     data = args.get("data", {})
-    print(f"[A2UI-DEBUG]   components={len(components)} data_keys={list(data.keys()) if data else []} surface={surface_id}")
 
     ops = [
         a2ui.create_surface(surface_id, catalog_id=catalog_id),
@@ -279,14 +226,12 @@ def generate_a2ui(runtime: ToolRuntime[Any]) -> str:
     if data:
         ops.append(a2ui.update_data_model(surface_id, data))
 
-    result = a2ui.render(operations=ops)
-    print(f"[A2UI-DEBUG] generate_a2ui DONE at t={time.time()-t0:.1f}s result_len={len(result)}")
-    return result
+    return a2ui.render(operations=ops)
 
 
 # ─── Graph ──────────────────────────────────────────────────────────
 
-model = ChatOpenAI(model="gpt-5.4", model_kwargs={"parallel_tool_calls": False})
+model = ChatOpenAI(model="gpt-5-mini", model_kwargs={"parallel_tool_calls": False})
 
 agent = create_agent(
     model=model,
@@ -305,9 +250,10 @@ agent = create_agent(
         - Flights: call search_flights to show flight cards with a pre-built schema.
         - Dashboards & rich UI: call generate_a2ui to create dashboard UIs with metrics,
           charts, tables, and cards. It handles rendering automatically.
-        - Charts: call query_data first, then render via pieChart or barChart (Controlled
-          Generative UI components registered on the frontend).
-        - Todos: call manage_todos to add, update, or remove todos.
+        - Charts: call query_data first, then render with the chart component.
+        - Todos: enable app mode first, then manage todos.
+        - A2UI actions: when you see a log_a2ui_event result (e.g. "view_details"),
+          respond with a brief confirmation. The UI already updated on the frontend.
     """,
 )
