+++ showcase/integrations/langgraph-python/src/app/demos/beautiful-chat/components/example-layout/index.tsx
@@ -1,8 +1,9 @@
 "use client";
 
-import { ReactNode, useState } from "react";
+import type { ReactNode } from "react";
+import { useState } from "react";
 import { ModeToggle } from "./mode-toggle";
-import { useFrontendTool } from "@copilotkit/react-core";
+import { useFrontendTool } from "@copilotkit/react-core/v2";
 
 interface ExampleLayoutProps {
   chatContent: ReactNode;
@@ -30,22 +31,23 @@ export function ExampleLayout({ chatContent, appContent }: ExampleLayoutProps) {
   });
 
   return (
-    <div className="h-full flex flex-row">
+    <div className="h-full flex flex-row pb-6">
       <ModeToggle mode={mode} onModeChange={setMode} />
 
       {/* Chat Content */}
       <div
-        className={`max-h-full flex flex-col ${
+        className={`max-h-full flex flex-col dark:bg-stone-950 ${
           mode === "app"
             ? "w-1/3 px-6 max-lg:hidden" // Hide on mobile in app mode
             : "flex-1 max-lg:px-4"
         }`}
       >
-        <div className="shrink-0 pt-6 pl-6 pb-2 max-lg:pl-4 max-lg:pt-4">
+        <div className="shrink-0 pt-6 pl-6 pb-2 max-lg:pl-4 max-lg:pt-4 flex gap-1.5 items-center align-center">
+          <span className="font-extrabold text-2xl pb-1.5">CopilotKit</span>
           <img
-            src="/copilotkit-logo.svg"
+            src="/copilotkit-logo-mark.svg"
             alt="CopilotKit"
-            className="h-7 dark:invert"
+            className="h-7"
           />
         </div>
         <div className="flex-1 min-h-0 overflow-y-auto">{chatContent}</div>
