+++ showcase/integrations/langgraph-python/src/app/demos/beautiful-chat/components/generative-ui/charts/bar-chart.tsx
@@ -138,9 +138,8 @@ export function BarChart({ title, description, data }: BarChartProps) {
               dataKey="value"
               radius={[6, 6, 0, 0]}
               maxBarSize={48}
-              // eslint-disable-next-line @typescript-eslint/no-explicit-any
               shape={
-                ((props: any) => (
+                ((props: Record<string, unknown>) => (
                   <AnimatedBar
                     {...props}
                     isNew={isNew(props.index as number)}
