+++ showcase/integrations/langgraph-python/src/app/demos/beautiful-chat/home-page.tsx
@@ -1,0 +1,23 @@
+"use client";
+
+import { CopilotChat } from "@copilotkit/react-core/v2";
+import { ExampleLayout } from "./components/example-layout";
+import { ExampleCanvas } from "./components/example-canvas";
+import { useGenerativeUIExamples, useExampleSuggestions } from "./hooks";
+
+export function HomePage() {
+  useGenerativeUIExamples();
+  useExampleSuggestions();
+
+  return (
+    <ExampleLayout
+      chatContent={
+        <CopilotChat
+          attachments={{ enabled: true }}
+          input={{ disclaimer: () => null, className: "pb-6" }}
+        />
+      }
+      appContent={<ExampleCanvas />}
+    />
+  );
+}
