+++ showcase/integrations/langgraph-python/src/app/demos/multimodal/page.tsx
@@ -23,265 +23,17 @@
  *   `<input type="file">` the paperclip path uses (DataTransfer + dispatch
  *   `change`). This keeps the sample and real-upload paths on a single
  *   code path — whatever works for one works for both.
- *
- * Legacy-shape rewrite:
- * - The published `@ag-ui/langgraph` converter (0.0.x) only understands
- *   the legacy `{ type: "binary", mimeType, data | url }` content-part
- *   shape when forwarding AG-UI messages to LangChain. The modern
- *   `{ type: "image" | "document", source: { type: "data" | "url", ... } }`
- *   parts that CopilotChat emits are silently dropped. Until the runtime
- *   ships an updated converter, we rewrite the outgoing user message in
- *   `onRunInitialized` so image/document/audio/video parts round-trip
- *   through the legacy shape the converter preserves.
  */
 
-import { useCallback, useEffect, useMemo } from "react";
-import { CopilotKit, CopilotChat, useAgent } from "@copilotkit/react-core/v2";
-import type { AttachmentUploadResult } from "@copilotkit/shared";
-
-import { SampleAttachmentButtons } from "./sample-attachment-buttons";
-
-/**
- * Minimal structural shape of an AG-UI message. `@ag-ui/client`'s
- * exported `Message` is a tagged union by role, but for the rewrite
- * we only care about the `user` branch and treat every other role as
- * pass-through. Keeping the type local avoids pulling `@ag-ui/client`
- * into this package's direct dependencies.
- */
-type AgentMessage = {
-  id?: string;
-  role: string;
-  content?: unknown;
-};
-
-/**
- * `onUpload` must resolve to an `AttachmentUploadResult` (data or url). We
- * always return the `data` variant — the demo inlines base64 instead of
- * uploading to external storage, matching the Wave 2b spec.
- */
-type DataUploadResult = Extract<AttachmentUploadResult, { type: "data" }>;
-
-const MAX_FILE_SIZE_BYTES = 10 * 1024 * 1024; // 10 MB
-const ACCEPT_MIME = "image/*,application/pdf";
-/**
- * Selector used by <SampleAttachmentButtons /> to locate CopilotChat's
- * hidden file input. Kept as a constant so the wrapper element and the
- * sample buttons cannot drift.
- */
-const CHAT_ROOT_SELECTOR = "[data-multimodal-demo-chat-root]";
-
-/**
- * Convert a File into the `AttachmentsConfig.onUpload` result shape —
- * inline base64 with the browser-provided mime type. We do this in the
- * browser rather than uploading to external storage because Wave 2b is a
- * self-contained demo; `maxSize: 10 MB` (set below) caps bloat.
- *
- * `FileReader` produces a `data:<mime>;base64,<payload>` URL; we strip the
- * prefix so the runtime forwards the raw base64 value (what the agent
- * expects in `source.value`).
- */
-function fileToDataAttachment(file: File): Promise<DataUploadResult> {
-  return new Promise((resolve, reject) => {
-    const reader = new FileReader();
-    reader.onerror = () =>
-      reject(reader.error ?? new Error(`FileReader failed for ${file.name}`));
-    reader.onload = () => {
-      const result = reader.result;
-      if (typeof result !== "string") {
-        reject(new Error(`Unexpected FileReader result type for ${file.name}`));
-        return;
-      }
-      // result looks like "data:image/png;base64,iVBORw0K..." — strip the prefix.
-      const commaIdx = result.indexOf(",");
-      const base64 = commaIdx >= 0 ? result.slice(commaIdx + 1) : result;
-      resolve({
-        type: "data",
-        value: base64,
-        mimeType: file.type || "application/octet-stream",
-        metadata: {
-          filename: file.name,
-          size: file.size,
-        },
-      });
-    };
-    reader.readAsDataURL(file);
-  });
-}
-
-/**
- * Rewrites a single `InputContent` part from the modern multimodal shape
- * (`type: "image" | "document" | "audio" | "video"`, `source.{type,value,mimeType}`)
- * to the legacy binary shape (`type: "binary"`, `mimeType`, `data` | `url`).
- *
- * The published `@ag-ui/langgraph` converter (see `aguiMessagesToLangChain`
- * in that package) only recognizes legacy `binary` parts — modern parts
- * are silently filtered out, so the LangGraph agent never sees them.
- * Returning the legacy shape keeps the attachment visible to the agent
- * while leaving everything else (CopilotChat UI, upload pipeline, etc.)
- * untouched. Text parts and already-legacy parts pass through unchanged.
- */
-function rewriteMultimodalPart(part: unknown): unknown {
-  if (!part || typeof part !== "object") return part;
-  const candidate = part as {
-    type?: string;
-    text?: string;
-    source?: {
-      type?: string;
-      value?: string;
-      mimeType?: string;
-    };
-  };
-  const type = candidate.type;
-  if (
-    type !== "image" &&
-    type !== "document" &&
-    type !== "audio" &&
-    type !== "video"
-  ) {
-    return part;
-  }
-  const source = candidate.source;
-  if (!source || typeof source.value !== "string") {
-    return part;
-  }
-  const mimeType = source.mimeType ?? "application/octet-stream";
-  if (source.type === "data") {
-    return {
-      type: "binary",
-      mimeType,
-      data: source.value,
-    };
-  }
-  if (source.type === "url") {
-    return {
-      type: "binary",
-      mimeType,
-      url: source.value,
-    };
-  }
-  return part;
-}
-
-/**
- * Walks a message list and rewrites user-message multimodal content parts
- * to the legacy `binary` shape. Non-user messages and plain-string user
- * messages pass through untouched. Idempotent: already-legacy `binary`
- * parts are a no-op for `rewriteMultimodalPart`.
- *
- * Returns the same array reference if nothing changed so the subscriber
- * in `onRunInitialized` can skip an unnecessary state write.
- */
-function rewriteMessagesForLegacyConverter(
-  messages: ReadonlyArray<Readonly<AgentMessage>>,
-): AgentMessage[] | null {
-  let mutated = false;
-  const next = messages.map((message) => {
-    if (message.role !== "user") return message as AgentMessage;
-    const content = message.content;
-    if (!Array.isArray(content)) return message as AgentMessage;
-    let partMutated = false;
-    const rewrittenParts = content.map((part) => {
-      const rewritten = rewriteMultimodalPart(part);
-      if (rewritten !== part) partMutated = true;
-      return rewritten;
-    });
-    if (!partMutated) return message as AgentMessage;
-    mutated = true;
-    return {
-      ...(message as object),
-      content: rewrittenParts,
-    } as AgentMessage;
-  });
-  return mutated ? next : null;
-}
-
-/**
- * Installs the `onRunInitialized` subscriber on the active agent. Scoped
- * to a small inner component so it can use the `useAgent` hook the
- * <CopilotChat> parent already relies on — subscribing there means we
- * rewrite messages on the *same* agent instance CopilotChat dispatches
- * through, even when threads are cloned or swapped.
- */
-function LegacyConverterShim() {
-  const { agent } = useAgent({ agentId: "multimodal-demo" });
-
-  const subscriber = useMemo(
-    () => ({
-      onRunInitialized: ({
-        messages,
-      }: {
-        messages: ReadonlyArray<Readonly<AgentMessage>>;
-      }) => {
-        const rewritten = rewriteMessagesForLegacyConverter(messages);
-        if (!rewritten) return;
-        return { messages: rewritten };
-      },
-    }),
-    [],
-  );
-
-  useEffect(() => {
-    if (!agent) return;
-    // AG-UI's AgentSubscriber type is tagged by role; our shim uses a
-    // structural message shape and returns only partial mutations, so
-    // cast at the subscribe boundary. The cast is safe: our
-    // onRunInitialized only ever returns a messages-mutation or void.
-    const handle = agent.subscribe(
-      subscriber as unknown as Parameters<typeof agent.subscribe>[0],
-    );
-    return () => handle.unsubscribe();
-  }, [agent, subscriber]);
-
-  return null;
-}
+import { CopilotKit } from "@copilotkit/react-core/v2";
+import { LegacyConverterShim } from "./legacy-converter-shim";
+import { MultimodalChat } from "./multimodal-chat";
 
 export default function MultimodalDemoPage() {
-  // `onUpload` is passed into CopilotChat's `AttachmentsConfig`. Both the
-  // paperclip button and the sample-injection path route files through
-  // this same function (sample buttons drive CopilotChat's hidden file
-  // input, which calls this internally via `useAttachments`). No
-  // duplicated upload code lives in the sample-button component.
-  const onUpload = useCallback(fileToDataAttachment, []);
-
   return (
     <CopilotKit runtimeUrl="/api/copilotkit-multimodal" agent="multimodal-demo">
       <LegacyConverterShim />
-      <div
-        data-testid="multimodal-demo-root"
-        className="mx-auto flex h-screen max-w-4xl flex-col gap-3 p-4 sm:p-6"
-      >
-        <header className="space-y-1">
-          <h1 className="text-lg font-semibold">Multimodal attachments</h1>
-          <p className="text-sm text-black/70 dark:text-white/70">
-            Attach an image or PDF with the paperclip, drag-and-drop onto the
-            chat, paste from the clipboard, or try one of the bundled samples
-            below. Then ask a question about the attachment and press send.
-          </p>
-        </header>
-
-        <SampleAttachmentButtons rootSelector={CHAT_ROOT_SELECTOR} />
-
-        <div
-          data-multimodal-demo-chat-root
-          className="min-h-0 flex-1 overflow-hidden rounded-lg border border-black/10 dark:border-white/10"
-        >
-          <CopilotChat
-            agentId="multimodal-demo"
-            className="h-full"
-            attachments={{
-              enabled: true,
-              accept: ACCEPT_MIME,
-              maxSize: MAX_FILE_SIZE_BYTES,
-              onUpload,
-              onUploadFailed: (err) => {
-                // Log without disrupting the default UI — CopilotChat already
-                // shows a toast-style indicator on validation failure.
-                console.warn("[multimodal-demo] attachment rejected", err);
-              },
-            }}
-          />
-        </div>
-      </div>
+      <MultimodalChat />
     </CopilotKit>
   );
 }
